When deleting repos, mark all other repos for fresh re-fetch.
TODO: This should be factored out as a method, and apparently called
also when enabling/disabling individual repos.

Consider following scenario: there's repo A with ETag support. Now
there's repo B which shadows some APK from repo A. If repo B then deleted,
info for the APK will be gone, and A's verion won't be fetched because of
ETag caching. So, we need to force re-fetching indexes from all repos
in such case.